refactor: 拆分 OneBot 分派测试模块
